Rename EntitiesContainer module to the components name
